+++ jeap-messaging-sequential-inbox-rest-api/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/api/config/SequentialInboxRestApiAutoConfiguration.java
@@ -1,0 +1,11 @@
+package ch.admin.bit.jeap.messaging.sequentialinbox.api.config;
+
+import ch.admin.bit.jeap.messaging.sequentialinbox.api.SequentialInboxController;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@AutoConfiguration
+@ComponentScan(basePackageClasses =  SequentialInboxController.class)
+class SequentialInboxRestApiAutoConfiguration {
+
+}